dev/core#495 - Administration pages field labels not getting translated
I'm guessing this is the right line to add this but I'm not set up to test it. See https://lab.civicrm.org/dev/core/issues/495 and https://civicrm.stackexchange.com/questions/34173/first-colums-of-form-not-translated. Also note that many of the files in the settings folder don't use ts() but that seems like an oversight since some do, but they aren't getting scanned.